requirements.txt
django
gunicorn
django-heroku
django-crispy-forms
pillow
django-import-export
